change QueryAstContext to CypherAstContext
