+++ src/planner/match/CypherClausePlanner.h
@@ -7,7 +7,7 @@
 #ifndef PLANNER_MATCH_CYPHERCLAUSEPLANNER_H_
 #define PLANNER_MATCH_CYPHERCLAUSEPLANNER_H_
 
-#include "context/ast/QueryAstContext.h"
+#include "context/ast/CypherAstContext.h"
 #include "planner/Planner.h"
 
 namespace nebula {
