+++ src/planner/match/MatchClausePlanner.cpp
@@ -6,7 +6,7 @@
 
 #include "planner/match/MatchClausePlanner.h"
 
-#include "context/ast/QueryAstContext.h"
+#include "context/ast/CypherAstContext.h"
 #include "planner/Query.h"
 #include "planner/match/Expand.h"
 #include "planner/match/MatchSolver.h"
