+++ src/planner/match/MatchPlanner.h
@@ -7,7 +7,7 @@
 #ifndef PLANNER_MATCHPLANNER_H_
 #define PLANNER_MATCHPLANNER_H_
 
-#include "context/ast/QueryAstContext.h"
+#include "context/ast/CypherAstContext.h"
 #include "planner/Planner.h"
 
 namespace nebula {
