+++ src/planner/match/SegmentsConnector.h
@@ -8,7 +8,7 @@
 #define PLANNER_MATCH_SEGMENTSCONNECTOR_H_
 
 #include "context/QueryContext.h"
-#include "context/ast/QueryAstContext.h"
+#include "context/ast/CypherAstContext.h"
 #include "planner/PlanNode.h"
 #include "planner/Planner.h"
 #include "planner/match/InnerJoinStrategy.h"
