+++ src/planner/match/StartVidFinder.h
@@ -7,7 +7,7 @@
 #ifndef PLANNER_MATCH_STARTVIDFINDER_H_
 #define PLANNER_MATCH_STARTVIDFINDER_H_
 
-#include "context/ast/QueryAstContext.h"
+#include "context/ast/CypherAstContext.h"
 #include "planner/Planner.h"
 
 namespace nebula {
